+++ errors.go
@@ -8,3 +8,11 @@ var (
 	errorCopyingSourceCode   = errors.New("error copying source code")
 	errorNoPackagesUpdatable = errors.New("no packages can be updated")
 )
+
+type errPackageNotFound struct {
+	path string
+}
+
+func (e errPackageNotFound) Error() string {
+	return "Package (" + e.path + ") not found"
+}