+++ rewrite.go
@@ -21,6 +21,7 @@ import (
 // according to the rules for func qualify.
 func rewrite(pkgs []*Package, qual string, paths []string) error {
 	for _, path := range pkgFiles(pkgs) {
+		debugln("rewrite", path)
 		err := rewriteTree(path, qual, paths)
 		if err != nil {
 			return err
@@ -51,17 +52,14 @@ func rewriteTree(path, qual string, paths []string) error {
 			continue
 		}
 		s := w.Stat()
-		switch s.IsDir() {
-		case true:
-			if s.Name() == "testdata" {
-				w.SkipDir()
-			}
-		case false:
-			if strings.HasSuffix(w.Path(), ".go") {
-				err := rewriteGoFile(w.Path(), qual, paths)
-				if err != nil {
-					return err
-				}
+		if s.IsDir() && s.Name() == "testdata" {
+			w.SkipDir()
+			continue
+		}
+		if strings.HasSuffix(w.Path(), ".go") {
+			err := rewriteGoFile(w.Path(), qual, paths)
+			if err != nil {
+				return err
 			}
 		}
 	}
@@ -71,6 +69,7 @@ func rewriteTree(path, qual string, paths []string) error {
 // rewriteGoFile rewrites import statments in the named file
 // according to the rules for func qualify.
 func rewriteGoFile(name, qual string, paths []string) error {
+	debugln("rewriteGoFile", name, ",", qual, ",", paths)
 	printerConfig := &printer.Config{Mode: printer.TabIndent | printer.UseSpaces, Tabwidth: 8}
 	fset := token.NewFileSet()
 	f, err := parser.ParseFile(fset, name, nil, parser.ParseComments)
@@ -117,14 +116,6 @@ func rewriteGoFile(name, qual string, paths []string) error {
 	}
 	return os.Rename(tpath, name)
 }
-
-// VendorExperiment is the Go 1.5 vendor directory experiment flag, see
-// https://github.com/golang/go/commit/183cc0cd41f06f83cb7a2490a499e3f9101befff
-var VendorExperiment = os.Getenv("GO15VENDOREXPERIMENT") == "1"
-
-// sep is the signature set of path elements that
-// precede the original path of an imported package.
-var sep = defaultSep(VendorExperiment)
 
 func defaultSep(experiment bool) string {
 	if experiment {